// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const IngredientType = {
  SYRUP: 'SYRUP',
  TOPPING: 'TOPPING'
} as const

export type IngredientType = (typeof IngredientType)[keyof typeof IngredientType]


export const OrderStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]